+++ metadata.rb
@@ -10,6 +10,7 @@ long_description 'Automates installation of Mac App Store apps'
 version          '0.1.1'
 
 depends          'build-essential', '~> 2.1'
+depends          'macosx_accessibility', '~> 0.1'
 
 supports         'mac_os_x'
 # rubocop:enable SingleSpaceBeforeFirstArg