+++ spec/spec_helper.rb
@@ -10,6 +10,8 @@ require 'coveralls'
 require 'tmpdir'
 require 'fileutils'
 require_relative '../libraries/matchers'
+require_relative 'support/resource/macosx_accessibility'
+require_relative 'support/provider/macosx_accessibility'
 
 RSpec.configure do |c|
   c.color = true