+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Mac App Store Cookbook CHANGELOG
 
 v?.?.? (????-??-??)
 -------------------
+- Remove the `mac_app_store_trusted_app` resource--it doesn't belong here
 
 v0.1.0 (2015-04-20)
 -------------------